# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectRequest": {
        "properties": {
          "region": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider data center; the default if omitted",
            "title": "Region"
          }
        },
        "title": "ConnectRequest",
        "type": "object"
      },
      "ConnectResponse": {
        "properties": {
          "redirect_url": {
            "description": "Send the browser here to continue (e.g. provider consent)",
            "title": "Redirect Url",
            "type": "string"
          }
        },
        "required": [
          "redirect_url"
        ],
        "title": "ConnectResponse",
        "type": "object"
      },
...
      "DatasetStatusOut": {
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "last_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "status": {
            "enum": [
              "ok",
              "failed",
              "never"
            ],
            "title": "Status",
            "type": "string"
          },
          "window_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "window_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          }
        },
        "required": [
          "key",
          "label",
          "status",
          "last_success_at",
          "last_attempt_at",
          "row_count",
          "window_start",
          "window_end",
          "error_code"
        ],
        "title": "DatasetStatusOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IntegrationList": {
        "properties": {
          "integrations": {
            "items": {
              "$ref": "#/components/schemas/IntegrationTile"
            },
            "title": "Integrations",
            "type": "array"
          }
        },
        "required": [
          "integrations"
        ],
        "title": "IntegrationList",
        "type": "object"
      },
      "IntegrationTile": {
        "properties": {
          "account_noun": {
            "description": "What one external account is called (\"organisation\")",
            "title": "Account Noun",
            "type": "string"
          },
          "availability": {
            "enum": [
              "available",
              "coming_soon"
            ],
            "title": "Availability",
            "type": "string"
          },
          "can_manage": {
            "description": "Whether the caller may connect/disconnect it",
            "title": "Can Manage",
            "type": "boolean"
          },
          "connection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectionOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "The live connection, if any"
          },
          "connection_state": {
            "enum": [
              "not_connected",
              "pending",
              "connected",
              "needs_reauth",
              "error",
              "coming_soon"
            ],
            "title": "Connection State",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "domain_name": {
            "title": "Domain Name",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "read_only": {
            "title": "Read Only",
            "type": "boolean"
          },
          "reads": {
            "description": "Plain-language list of the data Clario reads",
            "items": {
              "type": "string"
            },
            "title": "Reads",
            "type": "array"
          },
          "regions": {
            "description": "Data centers to choose from; first is default",
            "items": {
              "$ref": "#/components/schemas/RegionOut"
            },
            "title": "Regions",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "vendor": {
            "title": "Vendor",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "vendor",
          "domain",
          "domain_name",
          "summary",
          "reads",
          "read_only",
          "availability",
          "connection_state",
          "connection",
          "regions",
          "account_noun",
          "can_manage"
        ],
        "title": "IntegrationTile",
        "type": "object"
      },
...
      "MemberList": {
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/Member"
            },
            "title": "Members",
            "type": "array"
          }
        },
        "required": [
          "members"
        ],
        "title": "MemberList",
        "type": "object"
      },
...
      "Permission": {
        "enum": [
          "workspace.view",
          "finance.view",
          "assistant.use",
          "integrations.manage",
          "members.manage"
        ],
        "title": "Permission",
        "type": "string"
      },
...
      "RegionOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label"
        ],
        "title": "RegionOut",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "owner",
          "admin",
          "member",
          "viewer"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "SyncRequest": {
        "properties": {
          "mode": {
            "default": "manual",
            "description": "manual: 'Refresh now' (cooldown applies); if_stale: only when data is stale",
            "enum": [
              "manual",
              "if_stale"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "title": "SyncRequest",
        "type": "object"
      },
      "SyncRunOut": {
        "properties": {
          "api_calls": {
            "title": "Api Calls",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "succeeded",
              "partial",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "trigger": {
            "enum": [
              "initial",
              "manual",
              "stale",
              "scheduled"
            ],
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trigger",
          "status",
          "started_at",
          "finished_at",
          "api_calls",
          "error_code"
        ],
        "title": "SyncRunOut",
        "type": "object"
      },
      "SyncStatus": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Oldest successful refresh among the datasets; null until all have data",
            "title": "As Of"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetStatusOut"
            },
            "title": "Datasets",
            "type": "array"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncRunOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "The run in progress, else the latest one"
          },
          "state": {
            "enum": [
              "idle",
              "running"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "as_of",
          "run",
          "datasets"
        ],
        "title": "SyncStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceDetail": {
        "properties": {
          "base_currency": {
            "title": "Base Currency",
            "type": "string"
          },
          "fiscal_year_start_month": {
            "title": "Fiscal Year Start Month",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Permissions",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "status",
          "role",
          "timezone",
          "base_currency",
          "fiscal_year_start_month",
          "permissions"
        ],
        "title": "WorkspaceDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clario API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}": {
      "get": {
        "operationId": "get_workspace_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/connections/{connection_id}/sync": {
      "get": {
        "operationId": "get_sync_api_v1_workspaces__workspace_id__connections__connection_id__sync_get",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "How current this connection's data is",
        "tags": [
          "sync"
        ]
      },
      "post": {
        "operationId": "request_sync_api_v1_workspaces__workspace_id__connections__connection_id__sync_post",
        "parameters": [
          {
            "description": "Connection id",
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "description": "Connection id",
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SyncRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Not connected, or needs reconnecting"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Cooldown, or the provider's daily budget is nearly used up"
          }
        },
        "summary": "Refresh this connection's data (runs in the background)",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/integrations": {
      "get": {
        "operationId": "list_integrations_api_v1_workspaces__workspace_id__integrations_get",
        "parameters": [
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Integrations for this workspace",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/integrations/{integration}": {
      "get": {
        "operationId": "get_integration_api_v1_workspaces__workspace_id__integrations__integration__get",
        "parameters": [
          {
            "description": "Integration key, e.g. zoho-books",
            "in": "path",
            "name": "integration",
            "required": true,
            "schema": {
              "description": "Integration key, e.g. zoho-books",
              "maxLength": 64,
              "title": "Integration",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationTile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One integration",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/integrations/{integration}/connect": {
      "post": {
        "operationId": "connect_api_v1_workspaces__workspace_id__integrations__integration__connect_post",
        "parameters": [
          {
            "description": "Integration key, e.g. zoho-books",
            "in": "path",
            "name": "integration",
            "required": true,
            "schema": {
              "description": "Integration key, e.g. zoho-books",
              "maxLength": 64,
              "title": "Integration",
              "type": "string"
            }
          },
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ConnectRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Requires integrations.manage"
          },
          "404": {
            "description": "Unknown or not visible to this workspace"
          },
          "409": {
            "description": "Not available yet (coming soon)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start connecting (or reconnecting) an integration",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/members": {
      "get": {
        "operationId": "list_members_api_v1_workspaces__workspace_id__members_get",
        "parameters": [
          {
            "description": "Workspace id",
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "description": "Workspace id",
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace members",
        "tags": [
          "workspaces"
        ]
      }
    }
  }
}